Set up a Shiny app to use tarteaucitron

This function must be called from a Shiny app's UI in order for all other ShinyLemonPie functions to work. You can call useShinyLemonPie() from anywhere inside the UI, as long as the final app UI contains the result of useShinyLemonPie()

useShinyLemonPie(
  lang = "cn",
  orientation = "bottom",
  showAlertSmall = FALSE,
  cookieslist = FALSE,
  closePopup = FALSE,
  showIcon = TRUE,
  iconPosition = "BottomRight",
  iconSrc = "",
  DenyAllCta = TRUE,
  AcceptAllCta = TRUE,
  highPrivacy = TRUE,
  removeCredit = FALSE,
  moreInfoLink = TRUE,
  groupServices = FALSE,
  handleBrowserDNTRequest = FALSE,
  adblocker = FALSE,
  mandatory = TRUE,
  mandatoryCta = TRUE,
  hashtag = "#tarteaucitron",
  cookieName = "tarteaucitron",
  serviceDefaultState = "wait",
  privacyUrl = "",
  bodyPosition = "top",
  cookieDomain = "",
  readmoreLink = "",
  useExternalCss = FALSE,
  useExternalJs = FALSE,
  useBS5Css = FALSE
)

Arguments

lang

Language attribute for the ShinyLemonPie input. Available values are "ar","bg", "ca", "cn","cs", "da", "de", "el", "en", "es", "et", "fi", "fr", "hu", "it", "ja", "lt", "lv", "nl", "no", "oc, "pl", "pt", "ro", "ru", "se", "sk", "sv", "tr", "uk", "vi", "zh" for Arabic, Bulgarian, Catalan, Chinese, Czech, Danish, German, Greek, English, Spanish, Estonian, Finnish, French, Hungarian, Italian, Japanese, Lithuanian, Latvian, Dutch, Norwegian, Occitan, Polish, Portuguese, Romanian & Moldavian, Russian, Northern Sami, Slovak, Swedish, Turkish, Ukrainian, Vietnamese and Chinese respectively.

orientation

Banner position (top, bottom and middle)

showAlertSmall

logical. Show the small banner on bottom right (FALSE or TRUE)

cookieslist

logical. Show the cookie list (FALSE or TRUE)

closePopup

logical. Show a close X on the banner (FALSE or TRUE)

showIcon

Show cookie icon to manage cookies

iconPosition

Icon position (BottomRight, BottomLeft, TopRight and TopLeft)

iconSrc

Optional: path for icon image

DenyAllCta

logical. Show the deny all button (FALSE or TRUE)

AcceptAllCta

logical. Show the accept all button when highPrivacy is TRUE (FALSE or TRUE)

highPrivacy

logical. HIGHLY RECOMMANDED Disable auto consent (FALSE or TRUE)

removeCredit

logical. Remove credit link (FALSE or TRUE)

moreInfoLink

logical. Show more info link (FALSE or TRUE)

groupServices

logical. Group services by category (FALSE or TRUE)

handleBrowserDNTRequest

logical. If Do Not Track == 1, disallow all (FALSE or TRUE)

adblocker

logical. Show a Warning if an adblocker is detected (FALSE or TRUE)

mandatory

logical. Show a message about mandatory cookies (FALSE or TRUE)

mandatoryCta

logical. Show the disabled accept button when mandatory on (FALSE or TRUE)

hashtag

Open the panel with this hashtag (default : #tarteaucitron)

cookieName

Cookie name (default : "tarteaucitron")

serviceDefaultState

Default state (true, wait or false)

privacyUrl

Privacy policy url

bodyPosition

bottom or top to bring it as first element for accessibility

cookieDomain

Shared cookie for multisite ex : .my-multisite-domaine.fr

readmoreLink

Change the default read more link

useExternalCss

logical. If false, the tarteaucitron.css file will be loaded

useExternalJs

logical. If false, the tarteaucitron.js file will be loaded

useBS5Css

logical. If true, a css theme compatible with BS5 is loaded.

Value

Scripts that ShinyLemonPie requires that are automatically inserted to the app head tag. A side effect of calling this function is that a ShinyLemonPie directory is added as a resource path using shiny::addResourcePath().

Examples

useShinyLemonPie(
  lang = "en",
  orientation = "bottom",
  showAlertSmall = FALSE,
  cookieslist = FALSE,
  closePopup = FALSE,
  showIcon = TRUE,
  iconPosition = "BottomRight",
  iconSrc = "",
  DenyAllCta = TRUE,
  AcceptAllCta = TRUE,
  highPrivacy = TRUE,
  removeCredit = FALSE,
  moreInfoLink = TRUE,
  groupServices = FALSE,
  handleBrowserDNTRequest = FALSE,
  adblocker = FALSE,
  mandatory = TRUE,
  mandatoryCta = TRUE,
  hashtag = "#tarteaucitron",
  cookieName = "tarteaucitron",
  serviceDefaultState = "wait",
  privacyUrl = "",
  bodyPosition = "top",
  cookieDomain = "",
  readmoreLink = "",
  useExternalCss = FALSE,
  useExternalJs = FALSE,
  useBS5Css = FALSE
)
#>